requirements.txt
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
PyYAML==6.0.1
pytest==7.4.3
pytest-cov==4.1.0